+++ src/state.h
@@ -105,7 +105,7 @@ namespace StochHMM{
 		inline emm* getEmission(size_t iter){return emission[iter];};
 		
 		double get_emission_prob(sequences& seqs, size_t iter);  //get emission for given (position)
-		double get_transition_prob(sequences& seqs,int to ,int iter);  //get transition  (position,from or too)
+		double get_transition_prob(sequences& seqs, size_t to, size_t iter);  //get transition  (position,from or too)
 		double getEndTrans();
 		
 		void print();
